% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ang}
\alias{ang}
\title{Get angles of Cartesian coordinates}
\usage{
ang(x, y)
}
\arguments{
\item{x}{x values}

\item{y}{y values}
}
\value{
angles of (x,y)
}
\description{
Get angles of Cartesian coordinates
}
