% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_error}
\alias{calc_error}
\title{Calculate error of time prediction}
\usage{
calc_error(truth, pred)
}
\arguments{
\item{truth}{vector of true circadian times}

\item{pred}{vector of predicted circadian times}
}
\value{
List of
(1) data frame with the true time, predicted time, and differences
(2) root mean squared error
(3) accuracy, with correct defined as predicted within 2hrs of truth
}
\description{
Calculate error of time prediction
}
