% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_24hr}
\alias{convert_to_24hr}
\title{Converts time to [0,24)}
\usage{
convert_to_24hr(time)
}
\arguments{
\item{time}{vector of time}
}
\value{
vector of time values within 24 hours
}
\description{
Converts time to [0,24)
}
