% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_polar}
\alias{convert_to_polar}
\title{Converts Cartesian coordinates to polar}
\usage{
convert_to_polar(x, y)
}
\arguments{
\item{x}{x values}

\item{y}{y values}
}
\value{
data frame with columns angles, angular x and y coordinates
}
\description{
Converts Cartesian coordinates to polar
}
