% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_differences.R
\name{create_DiffMatrix}
\alias{create_DiffMatrix}
\title{Creates a matrix of gene expression  differences}
\usage{
create_DiffMatrix(genes, dat)
}
\arguments{
\item{genes}{vector of genes of interest, length G}

\item{dat}{data matrix with time x genes, size N x G}
}
\value{
matrix of gene expression differences, size N x G^2
}
\description{
Creates a matrix of gene expression  differences
}
