% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_differences.R
\name{filter_DiffMatrix}
\alias{filter_DiffMatrix}
\title{Filters difference matrix to keep only one pair of genes with differences > 0
Assumes effect of Gene1_Gene2 is not the same as Gene2_Gene1 on results}
\usage{
filter_DiffMatrix(diffs)
}
\arguments{
\item{diffs}{matrix of gene expression differences with time x gene pairs, size N x G^2}
}
\value{
matrix of gene expression differences, size N x ((G^2 - G)/2)
}
\description{
Filters difference matrix to keep only one pair of genes with differences > 0
Assumes effect of Gene1_Gene2 is not the same as Gene2_Gene1 on results
}
