% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhythmic_genes.R
\name{find_periodic_genes}
\alias{find_periodic_genes}
\title{Extracts the significant periodic genes from the meta2d output}
\usage{
find_periodic_genes(
  input_file,
  test_genes,
  per_method = c("LS", "JTK"),
  thres = 1.01
)
}
\arguments{
\item{input_file}{input file to read in, needs to be ".txt" (output of meta2d)}

\item{test_genes}{vector of genes in the testing data set}

\item{per_method}{vector of periodic method to use (LS or JTK)}

\item{thres}{threshold for p-value}
}
\description{
Extracts the significant periodic genes from the meta2d output
}
