% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda.R
\name{get_FDAcurves}
\alias{get_FDAcurves}
\title{Run FDA with a Fourier basis and positive constraints}
\usage{
get_FDAcurves(dat, time = NULL, numbasis = 5, seed_num = 12345)
}
\arguments{
\item{dat}{matrix of gene expressions with time x genes, size N x G}

\item{time}{vector of time}

\item{numbasis}{integer, number of Fourier basis}

\item{seed_num}{seed number}
}
\value{
matrix of smoothing curves, size M x G
}
\description{
Run FDA with a Fourier basis and positive constraints
}
