% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhythmic_genes.R
\name{run_metacycle}
\alias{run_metacycle}
\title{Runs the meta2d function from MetaCycle}
\usage{
run_metacycle(df, timepoints, out_file, out_dir, method = c("JTK", "LS"), ...)
}
\arguments{
\item{df}{expression data matrix with genes as rows and time/samples as columns}

\item{timepoints}{vector of time points, equal size to number of columns of df}

\item{out_file}{name of file in ".txt" format; saves expression matrix into text file for input into MetaCycle}

\item{out_dir}{output directory to save MetaCycle results}

\item{method}{character vector defining which method(s) to detect rhythmic signals}

\item{...}{additional parameters for meta2d from MetaCycle}
}
\description{
Runs the meta2d function from MetaCycle
}
