% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_differences.R
\name{scale_DiffMatrix}
\alias{scale_DiffMatrix}
\title{Scales difference matrix row-wise (by time) to be between 0 and 1}
\usage{
scale_DiffMatrix(diffs)
}
\arguments{
\item{diffs}{matrix of gene expression differences with time x gene pairs, size N x P}
}
\value{
matrix of gene expression differences, size N x P
}
\description{
Scales difference matrix row-wise (by time) to be between 0 and 1
}
