% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauFisher.R
\name{test_tauFisher}
\alias{test_tauFisher}
\title{Test tauFisher}
\usage{
test_tauFisher(chosen_genes, test_df, test_time, pca_embedding, trained_model)
}
\arguments{
\item{chosen_genes}{vector of chosen periodic genes to subset on}

\item{test_df}{gene expression matrix for training data (gene x sample/time)}

\item{test_time}{vector of times (e.g. c(0, 2, 4, ..., 26))}

\item{pca_embedding}{matrix of PCs and loadings from \code{train_tauFisher}}

\item{trained_model}{trained model from \code{train_tauFisher}}
}
\value{
data frame with true time and predicted time
}
\description{
Test tauFisher
}
