% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauFisher.R
\name{train_tauFisher}
\alias{train_tauFisher}
\title{Train tauFisher}
\usage{
train_tauFisher(periodic_genes, train_df, train_time, nrep = 1, numbasis = 5)
}
\arguments{
\item{periodic_genes}{vector of periodic genes to subset on}

\item{train_df}{gene expression matrix for training data (gene x sample/time)}

\item{train_time}{vector of time points (e.g. c(0, 4, 8, ..., 48, 52))}

\item{nrep}{number of replicates}

\item{numbasis}{number of basis functions for FDA}
}
\value{
list of (1) periodic genes in the training set,
(2) PCA embedding,
(3) regression model
}
\description{
Train tauFisher
}
